Solved issue with minimax, buttttt
Turns out that the issue was that I was getting the index in the wrong
list for the moves.
Now the other problem is that in my search for minimizing the amount of
validating I had to do for the final state (By just validating from the
last value entered) I ended up just considering as if the only
posibilities for victory were when the entered value was in a corner, so
for example if we had the following row: X| |X, and we entered the X in
the middle it would not register a win, since the current functions for
cheking only check either the left, or the right, not both a time (or at
least do something to simulate it). I am going to correct this. (This
happens with Columns and diagonals as well, not only Rows)